Fix (ci): Ensure `test-build` jobs precede `test-update` jobs to correctly test incremental builds (#112)
Previously, the `test-update` job ran independently of the `test-build` job. The `test-update` job was not an incremental build based on the `test-build` job.

Now, the two independent jobs are combined as one job, with `test-build` preceeding `test-update`. This properly tests incremental builds using the local docker cache. In order to prevent `test-update` from pulling a remote `:latest` image, but to use the `:latest` from the preceeding `test-build` step's build cache, `NO_PULL` is added for `PIPELINE=update`.

Co-authored-by: leo <leonardjonathanoh@live.com>
Co-authored-by: joe <joeltimothyoh@live.com>